Restructured UTs of extension application
