+++ broker/applications/extensions/test/acceptance/dashboard.docker.spec.js
@@ -1,26 +1,25 @@
 'use strict';
 
 const _ = require('lodash');
-const { CONST } = require('@sf/common-utils');
 const parseUrl = require('url').parse;
-const app = require('../support/apps').external;
+const app = require('../../../../test/test_broker/support/apps').external;
+const { CONST } = require('@sf/common-utils');
+const docker = require('@sf/docker');
+
 
 describe('dashboard', function () {
-  describe('virtualHost', function () {
+  describe('docker', function () {
 
-    const service_id = '19f17a7a-5247-4ee2-94b5-03eac6756388';
-    const plan_id = 'd035f948-5d3a-43d7-9aec-954e134c3e9d';
-    const plan_guid = 'f6280923-b144-4f02-adf7-76a7b5ef3a4a';
-    const instance_id = '5a877873-7659-40ea-bdcb-096e9ae0cbb3';
+    const service_id = '24731fb8-7b84-4f57-914f-c3d55d793dd4';
+    const plan_id = '466c5078-df6e-427d-8fb2-c76af50c0f56';
+    const instance_id = 'b3e03cb5-29cc-4fcf-9900-023cf149c554';
+    const service_plan_guid = '466c5078-df6e-427d-8fb2-c76af50c0f56';
     const organization_guid = 'b8cbbac8-6a20-42bc-b7db-47c205fccf9a';
     const space_guid = 'e7c0a437-7585-4d75-addf-aa4d45b49f3a';
-    const parent_instance_id = 'b4719e7c-e8d3-4f7f-c51c-769ad1c3ebfa';
-    const deployment_name = 'service-fabrik-0028-b4719e7c-e8d3-4f7f-c51c-769ad1c3ebfa';
-    const instance_name = 'rmq';
 
-    const resource1 = {
+    const resource = {
       apiVersion: 'deployment.servicefabrik.io/v1alpha1',
-      kind: 'VirtualHost',
+      kind: 'Docker',
       metadata: {
         name: instance_id,
         labels: {
@@ -39,13 +38,10 @@ describe('dashboard', function () {
           organization_guid: organization_guid,
           space_guid: space_guid,
           parameters: {
-            dedicated_rabbitmq_instance: `${instance_name}`
+            foo: 'bar'
           }
         })
       },
-      operatorMetadata: {
-        deploymentName: `${deployment_name}`
-      },
       status: {
         state: 'succeeded',
         lastOperation: '{}',
@@ -53,42 +49,20 @@ describe('dashboard', function () {
       }
     };
 
-    const resource2 = {
-      apiVersion: 'deployment.servicefabrik.io/v1alpha1',
-      kind: 'Director',
-      metadata: {
-        name: parent_instance_id,
-        labels: {
-          state: 'succeeded'
-        }
-      },
-      spec: {
-        options: JSON.stringify({
-          service_id: service_id,
-          plan_id: plan_id,
-          context: {
-            platform: 'cloudfoundry',
-            organization_guid: organization_guid,
-            space_guid: space_guid
-          },
-          organization_guid: organization_guid,
-          space_guid: space_guid
-        })
-      },
-      status: {
-        state: 'succeeded',
-        lastOperation: '{}',
-        response: '{}'
-      }
-    };
+    before(function () {
+      mocks.docker.getAllContainers([]);
+      return mocks.setup([
+        docker.updatePortRegistry()
+      ]);
+    });
+
+    afterEach(function () {
+      mocks.reset();
+    });
 
     describe('/manage/instances/:service_id/:plan_id/:instance_id', function () {
-
-      afterEach(function () {
-        mocks.reset();
-      });
-
       this.slow(1500);
+
       it('should redirect to authorization server', function () {
         const agent = chai.request.agent(app);
         agent.app.listen(0);
@@ -96,9 +70,14 @@ describe('dashboard', function () {
         mocks.uaa.getAccessTokenWithAuthorizationCode(service_id);
         mocks.uaa.getUserInfo();
         mocks.cloudController.getServiceInstancePermissions(instance_id);
-        mocks.apiServerEventMesh.nockGetResource(CONST.APISERVER.RESOURCE_GROUPS.DEPLOYMENT, CONST.APISERVER.RESOURCE_TYPES.VIRTUALHOST, instance_id, resource1, 2);
-        mocks.apiServerEventMesh.nockGetResource(CONST.APISERVER.RESOURCE_GROUPS.DEPLOYMENT, CONST.APISERVER.RESOURCE_TYPES.DIRECTOR, parent_instance_id, resource2, 1);
-        mocks.cloudController.findServicePlanByInstanceId(instance_id, plan_guid, plan_id, undefined, 2);
+        mocks.cloudController.findServicePlanByInstanceId(instance_id, service_plan_guid, plan_id, undefined, 2);
+        mocks.apiServerEventMesh.nockGetResource(CONST.APISERVER.RESOURCE_GROUPS.DEPLOYMENT, CONST.APISERVER.RESOURCE_TYPES.DOCKER, instance_id, resource, 1);
+        mocks.docker.inspectContainer(instance_id);
+        mocks.docker.inspectContainer(instance_id);
+        mocks.docker.inspectContainer();
+        mocks.docker.listContainerProcesses();
+        mocks.docker.getContainerLogs();
+        mocks.docker.inspectImage();
         return agent
           .get(`/manage/instances/${service_id}/${plan_id}/${instance_id}`)
           .set('Accept', 'application/json')
@@ -128,18 +107,16 @@ describe('dashboard', function () {
           .then(res => {
             expect(res.body.userId).to.equal('me');
             expect(res.body.instance.metadata.name).to.equal(instance_id);
-            expect(res.body.parent_instance.metadata.name).to.equal(parent_instance_id);
+            expect(res.body.processes).to.eql([
+              ['UID', 'PID'],
+              ['root', '13642']
+            ]);
             mocks.verify();
           });
       });
     });
 
-    describe('/manage/dashboards/virtual_host/instances/:instance_id', function () {
-
-      afterEach(function () {
-        mocks.reset();
-      });
-
+    describe('/manage/dashboards/docker/instances/:instance_id', function () {
       this.slow(1500);
       it('should redirect to authorization server', function () {
         const agent = chai.request.agent(app);
@@ -148,10 +125,12 @@ describe('dashboard', function () {
         mocks.uaa.getAccessTokenWithAuthorizationCode(service_id);
         mocks.uaa.getUserInfo();
         mocks.cloudController.getServiceInstancePermissions(instance_id);
-        mocks.apiServerEventMesh.nockGetResource(CONST.APISERVER.RESOURCE_GROUPS.DEPLOYMENT, CONST.APISERVER.RESOURCE_TYPES.VIRTUALHOST, instance_id, resource1, 3);
-        mocks.apiServerEventMesh.nockGetResource(CONST.APISERVER.RESOURCE_GROUPS.DEPLOYMENT, CONST.APISERVER.RESOURCE_TYPES.DIRECTOR, parent_instance_id, resource2, 1);
+        mocks.apiServerEventMesh.nockGetResource(CONST.APISERVER.RESOURCE_GROUPS.DEPLOYMENT, CONST.APISERVER.RESOURCE_TYPES.DOCKER, instance_id, resource, 3);
+        mocks.docker.inspectContainer(instance_id);
+        mocks.docker.listContainerProcesses();
+        mocks.docker.getContainerLogs();
         return agent
-          .get(`/manage/dashboards/virtual_host/instances/${instance_id}`)
+          .get(`/manage/dashboards/docker/instances/${instance_id}`)
           .set('Accept', 'application/json')
           .set('X-Forwarded-Proto', 'https')
           .redirects(2)
@@ -179,7 +158,10 @@ describe('dashboard', function () {
           .then(res => {
             expect(res.body.userId).to.equal('me');
             expect(res.body.instance.metadata.name).to.equal(instance_id);
-            expect(res.body.parent_instance.metadata.name).to.equal(parent_instance_id);
+            expect(res.body.processes).to.eql([
+              ['UID', 'PID'],
+              ['root', '13642']
+            ]);
             mocks.verify();
           });
       });
