+++ broker/applications/extensions/test/acceptance/service-fabrik-api-2.0.backups.director.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const app = require('../support/apps').external;
+const app = require('../../../../test/test_broker/support/apps').external;
 const config = require('@sf/app-config');
 const { CONST } = require('@sf/common-utils');
 const {
