+++ broker/applications/extensions/test/acceptance/service-fabrik-api-2.0.instances.director.spec.js
@@ -13,7 +13,7 @@ const {
     unifyDiffResult
   }
 } = require('@sf/common-utils');
-const apps = require('../support/apps');
+const apps = require('../../../../test/test_broker/support/apps');
 const {
   catalog,
   Service
